==> 1_DataAggregation.ipynb <==
# Data Aggregation

**Input:** streets_lat_long_curated.csv (containing latitue and longitude points from image donwloaded from crawler step

**Ouput:** geo.csv (contains the census sector corresponding to the lat-long points.


import pandas as pd
import numpy as np
import geopandas as gpd
import matplotlib.pyplot as plt
from shapely.geometry import Point, Polygon
import itertools

from tqdm.notebook import tqdm 

df = pd.read_csv('streets_lat_long_curated.csv')

df

df = df[df['pano_id'] != 'pano_id']
df['latitude'] = df['latitude'].astype(float)
df['longitude'] = df['longitude'].astype(float)

pano_counts = df['pano_id'].value_counts()
repeated_pano = pano_counts[pano_counts > 16].index
pano_counts[repeated_pano]

df[df['pano_id']=='amYzj4DUMQbgHzB9M3a_WA']['name'].value_counts()

df = df.drop_duplicates(subset=['pano_id', 'direction'])
df = df[(df['latitude'] < 10)&(df['longitude'] > -75)&(df['longitude'] < -25)]

df

geometry = [Point(xy) for xy in zip(df['longitude'].astype(float), df['latitude'].astype(float))]

world = gpd.read_file(gpd.datasets.get_path('naturalearth_lowres'))
brasil = world[world['iso_a3']=='BRA']

geo = gpd.GeoDataFrame(df, geometry=geometry)

def find_setor(row):
 
    for _, setor in vale.iterrows():
        if row['geometry'].within(setor['geometry']):
            return setor['CD_GEOCODI']
    else:
        return np.nan
    
geo['setor'] = geo.progress_apply(find_setor, axis=1)
#100%120772/120772 [3:09:21<00:00, 10.63it/s]

## saving data

#generated geo.csv
geo.to_csv('geo.csv', index=None)


count_setor = geo.groupby('setor')['uf'].count()

geo
